service_ai/requirements.txt
flask==2.0.2
numpy==1.22.0
pandas==1.3.5
scikit-surprise==1.1.1
